+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 Date format: DD/MM/YYYY
 
-## [4.0.0-pre.2] - Tabs and Bugs - [23/07/2022]
+## [4.0.0-pre.3] - [00/08/2022]
+
+- `NavigationView` top mode fixes:
+  - When top overflow menu is opened, `PaneItemHeader` no longer throws an unsupported error
+  - When on top mode, `PaneItemHeader` is properly aligned to the other items.
+  - Added `NavigationPaneThemeData.headerPadding`, which is applied to `PaneItemHeader` on open, compact and minimal mode. It defaults to 10 pixels at the top
+  - **BREAKING** `PaneItem.getPropertyFromTitle` is now `widget.getProperty`:
+
+  Before:
+  `getPropertyFromTitle<TextStyle>()`
+
+  Now:
+  `title.getProperty<TextStyle>()`
+
+  This was changed because the properties of `PaneItemHeader` needed to be accessed, but the old version only supported to get the properties of `PaneItem.title`. It can be called on a `Text`, `RichText` or in an `Icon` widget
+  - `InheritedNavigationView` is now accessible on the top overflow menu
+  - Added `NavigationPaneThemeData.selectedTopTextStyle` and `NavigationPaneThemeData.unselectedTopTextStyle`, which is applied to the items on top mode
+  - Updated default transitions for top mode: `HorizontalSlidePageTransition`
+
+## [4.0.0-pre.2] - Tabs, Tiles and Bugs - [23/07/2022]
 
 - Remove whitespace on `ContentDialog` if title is omitted ([#418](https://github.com/bdlukaa/fluent_ui/issues/418))
 - Apply correct color to the Date and Time Pickers button when selected ([#415](https://github.com/bdlukaa/fluent_ui/issues/415), [#417](https://github.com/bdlukaa/fluent_ui/issues/417))